+++ yrt-pet/include/yrt-pet/operators/SiddonKernels.cuh
@@ -186,13 +186,13 @@ __device__ void
 		int dir_z = (inv_p12_z >= 0.0) ? 1 : -1;
 
 		// 2. Intersection with volume
-		const int nx = imgParams.nx;
-		const int ny = imgParams.ny;
-		const int nz = imgParams.nz;
+		const size_t nx = imgParams.nx;
+		const size_t ny = imgParams.ny;
+		const size_t nz = imgParams.nz;
 		const float imgLength_x = imgParams.length_x;
 		const float imgLength_y = imgParams.length_y;
 		const float imgLength_z = imgParams.length_z;
-		const int num_xy = nx * ny;
+		const size_t num_xy = nx * ny;
 		const float dx = imgParams.vx;
 		const float dy = imgParams.vy;
 		const float dz = imgParams.vz;
@@ -240,24 +240,24 @@ __device__ void
 		float ax_next = flat_x ? FLT_MAX : ax_min;
 		if (!flat_x)
 		{
-			int kx =
-			    (int)ceil(dir_x * (a_cur * (p2.x - p1.x) - x_cur + p1.x) / dx);
+			size_t kx =
+			    ceil(dir_x * (a_cur * (p2.x - p1.x) - x_cur + p1.x) / dx);
 			x_cur += kx * dir_x * dx;
 			ax_next = (x_cur - p1.x) * inv_p12_x;
 		}
 		float ay_next = flat_y ? FLT_MAX : ay_min;
 		if (!flat_y)
 		{
-			int ky =
-			    (int)ceil(dir_y * (a_cur * (p2.y - p1.y) - y_cur + p1.y) / dy);
+			size_t ky =
+			    ceil(dir_y * (a_cur * (p2.y - p1.y) - y_cur + p1.y) / dy);
 			y_cur += ky * dir_y * dy;
 			ay_next = (y_cur - p1.y) * inv_p12_y;
 		}
 		float az_next = flat_z ? FLT_MAX : az_min;
 		if (!flat_z)
 		{
-			int kz =
-			    (int)ceil(dir_z * (a_cur * (p2.z - p1.z) - z_cur + p1.z) / dz);
+			size_t kz =
+			    ceil(dir_z * (a_cur * (p2.z - p1.z) - z_cur + p1.z) / dz);
 			z_cur += kz * dir_z * dz;
 			az_next = (z_cur - p1.z) * inv_p12_z;
 		}
@@ -333,12 +333,12 @@ __device__ void
 			}
 			if (!IsIncremental || flag_first)
 			{
-				vx = (int)((p1.x + a_mid * (p2.x - p1.x) + imgLength_x / 2) *
-				           inv_dx);
-				vy = (int)((p1.y + a_mid * (p2.y - p1.y) + imgLength_y / 2) *
-				           inv_dy);
-				vz = (int)((p1.z + a_mid * (p2.z - p1.z) + imgLength_z / 2) *
-				           inv_dz);
+				vx = static_cast<size_t>(
+				    (p1.x + a_mid * (p2.x - p1.x) + imgLength_x / 2) * inv_dx);
+				vy = static_cast<size_t>(
+				    (p1.y + a_mid * (p2.y - p1.y) + imgLength_y / 2) * inv_dy);
+				vz = static_cast<size_t>(
+				    (p1.z + a_mid * (p2.z - p1.z) + imgLength_z / 2) * inv_dz);
 				offset_img = vz * num_xy + vy * nx;
 				flag_first = false;
 				if (vx < 0 || vx >= nx || vy < 0 || vy >= ny || vz < 0 ||
